# Conflicts:
#	maps/Aurora-MapDev/aurora-2.dmm
#	maps/Aurora-MapDev/aurora-3.dmm
#	maps/Aurora-MapDev/aurora-5.dmm

+++ code/modules/turbolift/turbolift.dm
@@ -96,7 +96,7 @@ var/global/list/turbolifts = list()
 			if(istype(AM, /mob/living))
 				var/mob/living/M = AM
 				M.gib()
-			else if(AM.simulated)
+			else if(AM.simulated && !istype(AM, /mob/eye))
 				qdel(AM)
 
 	origin.move_contents_to(destination)
